Refactoring of LeadMe main
